+++ lib/eve_online.rb
@@ -51,6 +51,7 @@ require 'eve_online/esi/corporation'
 require 'eve_online/esi/corporation_industry_jobs'
 require 'eve_online/esi/corporation_blueprints'
 
+require 'eve_online/esi/alliances'
 require 'eve_online/esi/alliance'
 require 'eve_online/esi/alliance_icon'